Reveamped autofill contents
You used to be able to automatically fill a part of your template with a frontendModule by using {{container=name;autofill=module;data=some_data}}. This was confusing for a number of reasons: firstly we had to check for the autofill attribute in a number of places to do special handling of egde cases (which we sometimes forgot). Secondly, you had to name this container without this name being actually used anywhere.
The new way™ to do autofilling is just {{autofill=module;data=some_data}}. This will never show in any part of the admin area. The added bonus is nice as well: Use it with {{autofill=module;data=some_data;templateFlag=LEAVE_IDENTIFIERS}} to create additional (traditional, non-autofilled) containers that will be parsed and filled afterwards. It is planned to add the ability to forwards-declare those containers so that they can be filled in the admin area as well.