More possibilities for generic_frontend_module widget
generic_frontend_module widget can now be used to wrap any given template in a form and do widgetSave with the form’s serializeArrayKV. Unserialize logic is smart enough for most cases, though not yet for checkboxes and radio buttons. Inner widgets work, too (and are even waited for before trying to set the current value), just add the element and use data-widget-type. Perfect for small frontend modules who need to be quickly configurable but don’t require a preview.