New frontend module/edit widget structure
The saving and loading part is now delegated to a new language_object_control widget. Saving is no longer the responsibility of the individual frontend module. This means that instead of widgetSave, there is now the old getSaveData method (this time, though, getting passed the data instead of having to read from $_REQUEST).
Also added sensible default methods for widgetData and getWidget, making the legacy_frontend_module widget redundant.