Preserve row pointers during list reload
Using `connect_detail_with_list` used to save the current row the detail was editing.
Now, the idData of the row is preserved instead and the row is searched from there.
This makes it possible to have the list reloaded and still correctly register
single-row updates a detail widget has made.