fixed exception thrown by missing template_name
most probably it was the JS the missing template.is_default param that created this problem.
while fixing this i added an exception in case of wrong default template configuration.

A issue remains: the arg $bExcludeDefault = true
I do not know what this value is designed to do?
• this method is never called from any other file than page detail
• the default template is added anyway (at least if it is correctly configured)
I guess it can be removed