Inform user about inherited contents
We now have a section for info messages with each container that is displayed
when using the default page type in the page detail form.

Currently, there is one possible message, which is displayed if a container is empty
but does display content items inherited from an ancestor page.

Later, we might also inform the user about filling in containers marked declaration_only.
To do this, we’d add a data-attribute to the container and initialize the
message in `update_additional_container_info`.