The page_input widget may include virtual items
There is a new settings on the page_input widget, `include_virtual`. It is off by default (to be fully backwards-compatible).
When enabled, it includes not only navigation items from pages but also ones that are represented only virtually.

Selecting a virtual navigation item in the page_input widget will yield a value of the following form:

    <page_id>/path/to/virtual

This makes it relatively easy to store or retrieve a reference to the page and to store the additional path seperately.