Handle internal virtual links in richtext
RichtextUtil now takes the additional path leading to a virtual
subpage of a linked page and stored in the database as part of the identifier.
Retrieving the page is also taken care of. The correct links will be
output for backend as well as frontend.

The link’s format for backend output is /get_file/internal_link_proxy/<page_id>/path/to/virtual
with the last part being, not coincidentally, of the same format used
with the page_input widget when include_virtual is true.