Apply draggables to topmost elements only
Previously, when a draggable was released, all droppables underneath it were affected by the drop.
Now, only the topmost element is affected (even if not droppable).

Note: droppables that won’t be affected will still highlight due to technical constraints
but only the affected droppable will be highlighted after a successful drop.

See #153 for further information.